+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Test for computing abundance of taxonomic rank (used to create the barplot showing the coverage of EsMeCaTa predicitons on the samples).
 * Cytochrome c oxidase in other cycle to highlight aerobic taxa. Update associated template.
 * Creation of HMM functional profile showing the abundance of each HMM in the sample of the dataset.
+* Motif validation step from METABOLIC.
 
 ## Fix